added error handling and validation to controllers

+++ controllers/followController.js
@@ -1,14 +1,19 @@
 const { PrismaClient } = require("@prisma/client");
+const { validateUserID } = require("../utils/helpers") 
 
 const prisma = new PrismaClient();
 
 exports.follow = async (req, res) => {
   const followerId = req.decoded.userId;
   const followingId = Number(req.params.id);
+  validateUserID(followerId)
+  validateUserID(followingId)
 
   try {
     if (followerId === followingId) {
-      throw new Error("You can't follow yourself.");
+      return res.status(400).send({
+        error: "You can't follow yourself",
+      });
     }
     const newFollow = await prisma.user.update({
       where: {
@@ -28,9 +33,9 @@ exports.follow = async (req, res) => {
         },
       },
     });
-    res.json(newFollow);
+    res.status(201).json(newFollow);
   } catch (err) {
-    res.send({
+    res.status(500).send({
       error: `${err.message}`,
     });
   }
@@ -39,8 +44,15 @@ exports.follow = async (req, res) => {
 exports.unfollow = async (req, res) => {
   const followerId = req.decoded.userId;
   const followingId = Number(req.params.id);
+  validateUserID(followerId)
+  validateUserID(followingId)
 
   try {
+    if (followerId === followingId) {
+      return res.status(400).send({
+        error: "You can't unfollow yourself",
+      });
+    }
     const unfollow = await prisma.follow.delete({
       where: {
         followerId_followingId: {
@@ -49,9 +61,9 @@ exports.unfollow = async (req, res) => {
         },
       },
     });
-    res.json(unfollow);
+    res.status(200).json(unfollow);
   } catch (err) {
-    res.send({
+    res.status(500).send({
       error: `${err.message}`,
     });
   }
@@ -59,6 +71,8 @@ exports.unfollow = async (req, res) => {
 
 exports.allFollowers = async (req, res) => {
   const userId = Number(req.params.id);
+  validateUserID(followerId)
+  validateUserID(followingId)
 
   try {
     const followers = await prisma.user.findUnique({
@@ -69,9 +83,16 @@ exports.allFollowers = async (req, res) => {
         followedBy: true,
       },
     });
-    res.json(followers);
+
+    if (!followers) {
+      return res.status(404).send({
+        error: "User not found",
+      });
+    }
+
+    res.status(200).json(followers);
   } catch (err) {
-    res.send({
+    res.status(500).send({
       error: `${err.message}`,
     });
   }
